Allow API and KGI address to be configurable:
In the latest version of the Karlsen and Kaspa block explorer,
the REST-API server and the Graph Inspector address have been
preset in the source code. This setup complicates the use of
the explorer on a testnet. To address this, we have added two
new environment variables that allow for the specification of
these addresses when building the React frontend.

* REACT_APP_API_ADDRESS which is the public address of the
  REST API endpoint.
* REACT_APP_KGI_ADDRESS which is the public address of the
  graph inspector to inspect a specific block.

The API endpoint and Graph Inspector must operate on a web server
secured with SSL.